Customer Relation Management/CRM/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
